// src/app/models/libro.model.ts
import { DataCatalogo } from "./dataCatalogo.model";

export class Libro {

    idLibro?: number;
    titulo?: string;
    anio?: number;
    serie?: string;
    fechaRegistro?: Date;
    fechaActualizacion?: Date;
    estado?: number;
    categoriaLibro?: DataCatalogo;
    tipoLibro?: DataCatalogo;











}